chore: renames ResetableZoom to SnapBackZoom
