+++ src/index.ts
@@ -1,5 +1,5 @@
-export { default as ResetableZoom } from './components/ResetableZoom';
+export { default as SnapBackZoom } from './components/SnapBackZoom';
 export { useImageSize } from './hooks/useImageSize';
-export { calculateAspectRatioSize } from './utils';
+export { getAspectRatioSize } from './utils/getAspectRatioSize';
 
 export type * from './types';
